The user is able to create and delete his own blog posts

+++ includes/delete.php
@@ -5,14 +5,16 @@
 ?>
 
 <?php
+if (isset($_SESSION['isAdmin']) && $_SESSION['isAdmin'] == true) {
+    $id = $_GET['user_ID'];
+    $sql = "DELETE FROM `user` WHERE user_ID = $id";
+    $result = mysqli_query($conn, $sql);
 
-$id = $_GET['user_ID'];
-$sql = "DELETE FROM `user` WHERE user_ID = $id";
-$result = mysqli_query($conn, $sql);
-
-if ($result) {
-  header("Location: admin.php?msg=Data deleted successfully");
-} else {
-  echo "Failed: " . mysqli_error($conn);
+    if ($result) {
+        header("Location: admin.php?msg=Data deleted successfully");
+    } else {
+        echo "Failed: " . mysqli_error($conn);
+    }
+}else{
+    header("Location: admin.php?msg=You are not admin not allowed!");
 }
-
